rename driver db_interface to mongo_db
